Did a bunch of refactoring to dropbox and youtube funcs
Refactored the dropbox and youtube funcs to be easier to maintain and more error proof. Made it so the uploadFiles functions reads the files from the downloads directory, instead of generating the path with the title and extension. Also made downloadedFiles a map instead of a slice and added goroutines to uploadFiles